consumer/requirements.txt
pandas
kafka-python
boto3
requests
